chore: Move project and site models into separate files
