+++ terraso_backend/apps/project_management/models/__init__.py
@@ -1,0 +1,4 @@
+from .projects import Project
+from .sites import Site
+
+__all__ = ["Project", "Site"]
